requirements.txt
Pillow
matplotlib
mtcnn
tensorflow
scikit-learn
keras
keras.models
simple_image_download

